refactor(mcp): extract spec verification projection
